requirements.txt
Flask
mysql-connector-python
requests
Flask-WTF
Flask-Login
Werkzeug
seaborn
matplotlib